[fix] 홈 화면 캐시 및 동행모집글 바로가기 구현

+++ core/domain/src/main/kotlin/com/tripmate/android/domain/repository/MateRepository.kt
@@ -6,7 +6,7 @@ import com.tripmate.android.domain.entity.MateRecruitmentEntity
 interface MateRepository {
     suspend fun checkPersonalizationCompletion(): Boolean
     suspend fun completePersonalization(flag: Boolean)
-    suspend fun companionApply(companionId: Int): Result<Unit>
-    suspend fun getCompanionsDetailInfo(companionId: Int): Result<MateRecruitPostEntity>
+    suspend fun companionApply(companionId: Long): Result<Unit>
+    suspend fun getCompanionsDetailInfo(companionId: Long): Result<MateRecruitPostEntity>
     suspend fun createCompanionRecruitment(mateRecruitmentEntity: MateRecruitmentEntity): Result<Unit>
 }
