+++ feature/trip-detail/src/main/kotlin/com/tripmate/android/feature/trip_detail/viewmodel/TripDetailUiAction.kt
@@ -5,6 +5,6 @@ sealed interface TripDetailUiAction {
     data class OnTabChanged(val index: Int) : TripDetailUiAction
     data object GetTripDetailInfo : TripDetailUiAction
     data object OnClickMateRecruit : TripDetailUiAction
-    data class OnClickMateReviewPost(val companionId: Int) : TripDetailUiAction
+    data class OnClickMateReviewPost(val companionId: Long) : TripDetailUiAction
     data object OnClickReport : TripDetailUiAction
 }
